Fix build

+++ suripu-app/src/test/java/com/hello/suripu/app/utils/TrackerMotionDataSourceTest.java
@@ -24,12 +24,12 @@ public class TrackerMotionDataSourceTest {
         final DateTime now = DateTime.now();
         data.add(new TrackerMotion(0L, 0L, 0L, now.getMillis(), -1, DateTimeZone.getDefault().getOffset(now)));  // this one should be skipped
         data.add(new TrackerMotion(0L, 0L, 0L, now.getMillis(), -2, DateTimeZone.getDefault().getOffset(now)));  // -2 & 0xFFFFFFFF
-        assertThat(Long.valueOf(TrackerMotionDataSource.getMinAmplitude(data)), is(0L));
+        assertThat(Integer.valueOf(TrackerMotionDataSource.getMinAmplitude(data)), is(0));
 
         data.clear();
         data.add(new TrackerMotion(0L, 0L, 0L, now.getMillis(), -2, DateTimeZone.getDefault().getOffset(now)));  // -2 & 0xFFFFFFFF
         data.add(new TrackerMotion(0L, 0L, 0L, now.getMillis(), 1, DateTimeZone.getDefault().getOffset(now)));  // -2 & 0xFFFFFFFF
-        assertThat(TrackerMotionDataSource.getMinAmplitude(data), is(1L));
+        assertThat(TrackerMotionDataSource.getMinAmplitude(data), is(1));
 
     }
